refactor: 체스 게임이 Status와 협력하도록 수정
